pyproject.toml
[project]
name = "revancedbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fdroidserver>=2.4.2",
    "pygithub>=2.8.1",
]

[project.scripts]
repatcher = "revancedbot:run_patcher"

[tool.uv]
package = true